# src/main/resources/swagger.json
{
  "openapi": "3.0.3",
  "info": {
    "title": "ECasofond REST API",
    "description": "Endpoints for the ECasofond project",
    "version": "0.6.9"
  },
  "servers": [
    {
      "url": "http://localhost:8080",
      "description": "Local server"
    }
  ],
  "security": [
    {
      "basicAuth": [],
      "bearerAuth": []
    }
  ],
  "paths": {
...
    "/company/all": {
      "get": {
        "tags": [
          "Company"
        ],
        "summary": "Get all companies",
        "operationId": "getAllCompanies",
        "security": [
          {
            "bearerAuth": []
          }
        ],
        "responses": {
          "200": {
            "description": "OK",
            "content": {
              "application/json": {
                "schema": {
                  "type": "array",
                  "items": {
                    "$ref": "#/components/schemas/Company"
                  }
                }
              }
            }
          },
          "401": {
            "description": "Unauthorized"
          }
        }
      }
    },
    "/company/{id}": {
      "get": {
        "tags": [
          "Company"
        ],
        "summary": "Get company by id",
        "operationId": "getCompanyById",
        "security": [
          {
            "bearerAuth": []
          }
        ],
        "parameters": [
          {
            "$ref": "#/components/parameters/id"
          }
        ],
        "responses": {
          "200": {
            "description": "OK",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/Company"
                }
              }
            }
          },
          "401": {
            "description": "Unauthorized"
          },
          "404": {
            "description": "Not found"
          }
        }
      }
    },
    "/company/add": {
      "post": {
        "tags": [
          "Company"
        ],
        "summary": "Add company",
        "operationId": "addCompany",
        "security": [
          {
            "bearerAuth": []
          }
        ],
        "requestBody": {
          "$ref": "#/components/requestBodies/AddCompanyBody"
        },
        "responses": {
          "200": {
            "description": "OK"
          },
          "401": {
            "description": "Unauthorized"
          },
          "403": {
            "description": "Forbidden"
          },
          "409": {
            "description": "Conflict"
          }
        }
      }
    },
    "/company/update": {
      "put": {
        "tags": [
          "Company"
        ],
        "summary": "Update company",
        "operationId": "updateCompany",
        "security": [
          {
            "bearerAuth": []
          }
        ],
        "requestBody": {
          "$ref": "#/components/requestBodies/UpdateCompanyBody"
        },
        "responses": {
          "200": {
            "description": "OK"
          },
          "401": {
            "description": "Unauthorized"
          },
          "403": {
            "description": "Forbidden"
          },
          "404": {
            "description": "Not found"
          },
          "409": {
            "description": "Conflict"
          }
        }
      }
    },
    "/company/delete/{id}": {
      "delete": {
        "tags": [
          "Company"
        ],
        "summary": "Delete company",
        "operationId": "deleteCompany",
        "security": [
          {
            "bearerAuth": []
          }
        ],
        "parameters": [
          {
            "$ref": "#/components/parameters/id"
          }
        ],
        "responses": {
          "200": {
            "description": "OK"
          },
          "401": {
            "description": "Unauthorized"
          },
          "403": {
            "description": "Forbidden"
          },
          "404": {
            "description": "Not found"
          }
        }
      }
    },
...
    "/entrytype/all": {
      "get": {
        "tags": [
          "EntryType"
        ],
        "summary": "Get all entry types",
        "operationId": "getAllEntryTypes",
        "security": [
          {
            "bearerAuth": []
          }
        ],
        "responses": {
          "200": {
            "description": "OK",
            "content": {
              "application/json": {
                "schema": {
                  "type": "array",
                  "items": {
                    "$ref": "#/components/schemas/EntryType"
                  }
                }
              }
            }
          },
          "401": {
            "description": "Unauthorized"
          }
        }
      }
    },
    "/entrytype/{id}": {
      "get": {
        "tags": [
          "EntryType"
        ],
        "summary": "Get entry type by id",
        "operationId": "getEntryTypeById",
        "security": [
          {
            "bearerAuth": []
          }
        ],
        "parameters": [
          {
            "$ref": "#/components/parameters/id"
          }
        ],
        "responses": {
          "200": {
            "description": "OK",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/EntryType"
                }
              }
            }
          },
          "401": {
            "description": "Unauthorized"
          },
          "404": {
            "description": "Not found"
          }
        }
      }
    },
    "/entrytype/add": {
      "post": {
        "tags": [
          "EntryType"
        ],
        "summary": "Add entry type",
        "operationId": "addEntryType",
        "security": [
          {
            "bearerAuth": []
          }
        ],
        "requestBody": {
          "$ref": "#/components/requestBodies/AddEntryTypeBody"
        },
        "responses": {
          "200": {
            "description": "OK"
          },
          "401": {
            "description": "Unauthorized"
          },
          "403": {
            "description": "Forbidden"
          },
          "409": {
            "description": "Conflict"
          }
        }
      }
    },
    "/entrytype/update": {
      "put": {
        "tags": [
          "EntryType"
        ],
        "summary": "Update entry type",
        "operationId": "updateEntryType",
        "security": [
          {
            "bearerAuth": []
          }
        ],
        "requestBody": {
          "$ref": "#/components/requestBodies/UpdateEntryTypeBody"
        },
        "responses": {
          "200": {
            "description": "OK"
          },
          "401": {
            "description": "Unauthorized"
          },
          "403": {
            "description": "Forbidden"
          },
          "404": {
            "description": "Not found"
          },
          "409": {
            "description": "Conflict"
          }
        }
      }
    },
    "/entrytype/delete/{id}": {
      "delete": {
        "tags": [
          "EntryType"
        ],
        "summary": "Delete entry type",
        "operationId": "deleteEntryType",
        "security": [
          {
            "bearerAuth": []
          }
        ],
        "parameters": [
          {
            "$ref": "#/components/parameters/id"
          }
        ],
        "responses": {
          "200": {
            "description": "OK"
          },
          "401": {
            "description": "Unauthorized"
          },
          "403": {
            "description": "Forbidden"
          },
          "404": {
            "description": "Not found"
          }
        }
      }
    },
    "/entry/all": {
      "get": {
        "tags": [
          "Entry"
        ],
        "summary": "Get all entries",
        "operationId": "getAllEntries",
        "security": [
          {
            "bearerAuth": []
          }
        ],
        "responses": {
          "200": {
            "description": "OK",
            "content": {
              "application/json": {
                "schema": {
                  "type": "array",
                  "items": {
                    "$ref": "#/components/schemas/Entry"
                  }
                }
              }
            }
          },
          "401": {
            "description": "Unauthorized"
          }
        }
      }
    },
    "/entry/{id}": {
      "get": {
        "tags": [
          "Entry"
        ],
        "summary": "Get entry by id",
        "operationId": "getEntryById",
        "security": [
          {
            "bearerAuth": []
          }
        ],
        "parameters": [
          {
            "$ref": "#/components/parameters/id"
          }
        ],
        "responses": {
          "200": {
            "description": "OK",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/Entry"
                }
              }
            }
          },
          "401": {
            "description": "Unauthorized"
          },
          "404": {
            "description": "Not found"
          }
        }
      }
    },
    "/entry/add": {
      "post": {
        "tags": [
          "Entry"
        ],
        "summary": "Add entry",
        "operationId": "addEntry",
        "security": [
          {
            "bearerAuth": []
          }
        ],
        "requestBody": {
          "$ref": "#/components/requestBodies/AddEntryBody"
        },
        "responses": {
          "200": {
            "description": "OK"
          },
          "401": {
            "description": "Unauthorized"
          },
          "403": {
            "description": "Forbidden"
          },
          "409": {
            "description": "Conflict"
          }
        }
      }
    },
    "/entry/update": {
      "put": {
        "tags": [
          "Entry"
        ],
        "summary": "Update entry",
        "operationId": "updateEntry",
        "security": [
          {
            "bearerAuth": []
          }
        ],
        "requestBody": {
          "$ref": "#/components/requestBodies/UpdateEntryBody"
        },
        "responses": {
          "200": {
            "description": "OK"
          },
          "401": {
            "description": "Unauthorized"
          },
          "403": {
            "description": "Forbidden"
          },
          "404": {
            "description": "Not found"
          },
          "409": {
            "description": "Conflict"
          }
        }
      }
    },
    "/entry/delete/{id}": {
      "delete": {
        "tags": [
          "Entry"
        ],
        "summary": "Delete entry",
        "operationId": "deleteEntry",
        "security": [
          {
            "bearerAuth": []
          }
        ],
        "parameters": [
          {
            "$ref": "#/components/parameters/id"
          }
        ],
        "responses": {
          "200": {
            "description": "OK"
          },
          "401": {
            "description": "Unauthorized"
          },
          "403": {
            "description": "Forbidden"
          },
          "404": {
            "description": "Not found"
          }
        }
      }
    }
  },
  "components": {
    "securitySchemes": {
      "basicAuth": {
        "type": "http",
        "scheme": "basic"
      },
      "BearerAuth": {
        "type": "http",
        "scheme": "bearer",
        "bearerFormat": "JWT"
      }
    },
    "schemas": {
...
      "Company": {
        "type": "object",
        "properties": {
          "id": {
            "type": "integer"
          },
          "name": {
            "type": "string"
          }
        }
      },
      "Role": {
        "type": "object",
        "properties": {
          "id": {
            "type": "integer"
          },
          "name": {
            "type": "string"
          },
          "level": {
            "type": "integer"
          },
          "write": {
            "type": "boolean"
          }
        }
      },
      "Profession": {
        "type": "object",
        "properties": {
          "id": {
            "type": "integer"
          },
          "name": {
            "type": "string"
          }
        }
      },
      "EntryType": {
        "type": "object",
        "properties": {
          "id": {
            "type": "integer"
          },
          "name": {
            "type": "string"
          }
        }
      },
      "Entry": {
        "type": "object",
        "properties": {
          "id": {
            "type": "integer"
          },
          "user": {
            "$ref": "#/components/schemas/User"
          },
          "type": {
            "$ref": "#/components/schemas/EntryType"
          },
          "description": {
            "type": "string"
          },
          "hourCount": {
            "type": "integer"
          },
          "created": {
            "type": "string",
            "format": "date-time"
          },
          "updated": {
            "type": "string",
            "format": "date-time"
          }
        }
      },
      "User": {
        "type": "object",
        "properties": {
          "id": {
            "type": "integer"
          },
          "username": {
            "type": "string"
          },
          "email": {
            "type": "string"
          },
          "displayName": {
            "type": "string"
          },
          "company": {
            "$ref": "#/components/schemas/Company"
          },
          "role": {
            "$ref": "#/components/schemas/Role"
          },
          "profession": {
            "$ref": "#/components/schemas/Profession"
          }
        }
      }
    },
    "requestBodies": {
...
      "AddCompanyBody": {
        "content": {
          "application/json": {
            "schema": {
              "$ref": "#/components/schemas/Company"
            }
          }
        }
      },
      "UpdateCompanyBody": {
        "content": {
          "application/json": {
            "schema": {
              "$ref": "#/components/schemas/Company"
            }
          }
        }
      },
      "AddEntryBody": {
        "content": {
          "application/json": {
            "schema": {
              "$ref": "#/components/schemas/Entry"
            }
          }
        }
      },
      "UpdateEntryBody": {
        "content": {
          "application/json": {
            "schema": {
              "$ref": "#/components/schemas/Entry"
            }
          }
        }
      },
      "AddEntryTypeBody": {
        "content": {
          "application/json": {
            "schema": {
              "$ref": "#/components/schemas/EntryType"
            }
          }
        }
      },
      "UpdateEntryTypeBody": {
        "content": {
          "application/json": {
            "schema": {
              "$ref": "#/components/schemas/EntryType"
            }
          }
        }
      },
...
    },
    "parameters": {
      "id": {
        "name": "id",
        "in": "path",
        "description": "Id",
        "required": true,
        "schema": {
          "type": "integer"
        }
      }
    }
  }
}